solver options splitting for better readability.
